+++ includes/content-registration/custom-post-types-registration.php
@@ -501,12 +501,16 @@ function register_content_fields_with_graphql( TypeRegistry $type_registry ) {
 				$field['type'] = array( 'list_of' => 'String' );
 			}
 
-			if ( isset( $field['isRepeatable'] ) && $field['isRepeatable'] ) {
+			if ( isset( $field['isRepeatable'] ) && $field['isRepeatable'] && 'text' === $field['type'] ) {
 				$field['type'] = array( 'list_of' => 'String' );
 			}
 
 			if ( isset( $field['isRepeatable'] ) && $field['isRepeatable'] && 'MediaItem' === $field['type'] ) {
 				$field['type'] = array( 'list_of' => 'MediaItem' );
+			}
+
+			if ( isset( $field['isRepeatableNumber'] ) && $field['isRepeatableNumber'] && 'Float' === $field['type'] ) {
+				$field['type'] = array( 'list_of' => 'Float' );
 			}
 
 			$field['resolve'] = static function( Post $post, $args, $context, $info ) use ( $field, $rich_text ) {