+++ includes/publisher/js/src/components/Text/index.js
@@ -1,1 +1,1 @@
-export { default as Text } from "./Text";
+export { default } from "./Text";